requirements.txt
Django==4.2.10
djangorestframework==3.14.0
pandas>=2.2.2
openpyxl==3.1.2
xlrd >= 2.0.1
gunicorn>=20.1.0
numpy>=2.0.0
requests>=2.31.0
python-dotenv==1.0.0
django-cors-headers==3.14.0
django-environ==0.9.0
boto3>=1.34.0
PyJWT>=2.10.1
bcrypt>=4.3.0
